+++ sdk/storage/storage-blob/src/policies/StorageBearerTokenChallengeAuthenticationPolicy.ts
@@ -18,61 +18,6 @@ import { delay } from "@azure/core-http";
 const Constants = {
   DefaultScope: "/.default",
   /**
-   * The core-http version
-   */
-  coreHttpVersion: "2.2.5",
-
-  /**
-   * Specifies HTTP.
-   */
-  HTTP: "http:",
-
-  /**
-   * Specifies HTTPS.
-   */
-  HTTPS: "https:",
-
-  /**
-   * Specifies HTTP Proxy.
-   */
-  HTTP_PROXY: "HTTP_PROXY",
-
-  /**
-   * Specifies HTTPS Proxy.
-   */
-  HTTPS_PROXY: "HTTPS_PROXY",
-
-  /**
-   * Specifies NO Proxy.
-   */
-  NO_PROXY: "NO_PROXY",
-
-  /**
-   * Specifies ALL Proxy.
-   */
-  ALL_PROXY: "ALL_PROXY",
-
-  HttpConstants: {
-    /**
-     * Http Verbs
-     */
-    HttpVerbs: {
-      PUT: "PUT",
-      GET: "GET",
-      DELETE: "DELETE",
-      POST: "POST",
-      MERGE: "MERGE",
-      HEAD: "HEAD",
-      PATCH: "PATCH",
-    },
-
-    StatusCodes: {
-      TooManyRequests: 429,
-      ServiceUnavailable: 503,
-    },
-  },
-
-  /**
    * Defines constants for use with HTTP headers.
    */
   HeaderConstants: {
@@ -80,20 +25,6 @@ const Constants = {
      * The Authorization header.
      */
     AUTHORIZATION: "authorization",
-
-    AUTHORIZATION_SCHEME: "Bearer",
-
-    /**
-     * The Retry-After response-header field can be used with a 503 (Service
-     * Unavailable) or 349 (Too Many Requests) responses to indicate how long
-     * the service is expected to be unavailable to the requesting client.
-     */
-    RETRY_AFTER: "Retry-After",
-
-    /**
-     * The UserAgent header.
-     */
-    USER_AGENT: "User-Agent",
   },
 };
 
@@ -352,7 +283,7 @@ export function storageBearerTokenChallengeAuthenticationPolicy(
   scopes: string | string[]
 ): RequestPolicyFactory {
   // This simple function encapsulates the entire process of reliably retrieving the token
-  let getToken = createTokenCycler(credential, scopes /* , options */);
+  let getToken = createTokenCycler(credential, scopes);
 
   class StorageBearerTokenChallengeAuthenticationPolicy extends BaseRequestPolicy {
     public constructor(nextPolicy: RequestPolicy, options: RequestPolicyOptions) {
@@ -387,10 +318,7 @@ export function storageBearerTokenChallengeAuthenticationPolicy(
           const parsedAuthUri = URLBuilder.parse(challengeInfo.authorization_uri);
           const pathSegments = parsedAuthUri.getPath()!.split("/");
           const tenantId = pathSegments[1];
-          const getTokenForChallenge = createTokenCycler(
-            credential,
-            challengeScopes /* , options */
-          );
+          const getTokenForChallenge = createTokenCycler(credential, challengeScopes);
 
           const tokenForChallenge = (
             await getTokenForChallenge({